Invalid DateTime

+++ ps9/finance/app.py
@@ -76,7 +76,7 @@ def buy():
             shares = (request.form.get("shares"))
             stock = str(request.form.get("symbol"))
             stock = stock.upper()
-            total = float(price) * int(shares)
+            total = float(price) * float(shares)
             newcash = cash - int(total)
             if cash - total > 0:
                 current_datetime = datetime.now()
